firmware: implement status query

+++ blazing-fan-firmware/src/ports/rp2040_port.rs
@@ -2,7 +2,7 @@
 pub enum RP2040Error {}
 
 pub trait RP2040Port {
-    fn board_tmp(&mut self) -> Result<f32, RP2040Error>;
+    fn board_tmp(&mut self) -> Result<i8, RP2040Error>;
 
     fn board_sys_voltage(&mut self) -> Result<f32, RP2040Error>;
